feat: removed our awaitable and stoppable interfaces
I came the conclusion that the interfaces were just causing more work for minimal benefit.

Yes technically this is breaking but again, I'm the only user of this lib and I can;t be bothered with the whole branching / module renaming thing until I can get semantic release to automate the whole process.